requirements.txt
fastapi
uvicorn[standard]
python-dotenv
prometheus-client
